// srm_ui/src/app/main/Models/etendering/ViewModels/supplier-view-model.ts
export class SupplierViewModel {
    id: string;
    supplierName: string;
    supplierType: string;
    contactEmail: string;
    contactName: string;
    supplierStatus: string;
    ifsSupplierId: string;
    title: string;
    isChecked: boolean;
    supplierGroupId: string;
    contactTitle: string;
    invited: string;
    catagory: string;
    isContactSave: boolean;
    currencyId: string;
    invitedCount: number;
    sourcingEventsAwarded: number;
    firstName: string;
    lastName: string;
}